+++ weaver/utils.py
@@ -563,22 +563,6 @@ def now(tz_name=None):
     return localize_datetime(datetime.now().astimezone(), tz_name=tz_name)
 
 
-def repr_json(data, force_string=True, **kwargs):
-    # type: (Any, bool, Any) -> Union[JSON, str, None]
-    """
-    Ensure that the input data can be serialized as JSON to return it formatted representation as such.
-
-    If formatting as JSON fails, returns the data as string representation or ``None`` accordingly.
-    """
-    if data is None:
-        return None
-    try:
-        data_str = json.dumps(data, **kwargs)
-        return data_str if force_string else data
-    except Exception:  # noqa: W0703 # nosec: B110
-        return str(data)
-
-
 def wait_secs(run_step=-1):
     # type: (int) -> int
     """