+++ CHANGES.rst
@@ -14,6 +14,8 @@ Changes:
 --------
 - Add `CLI` option ``-L/--no-links`` that drops the ``links`` section of any response to make the printed result more
   concise and specific to relevant details of the called operation.
+- Add `CLI` option ``-F/--format`` that allows output of contents in an alternative format.
+  Available formatters include JSON, YAML and XML representations, with either pretty indentation and newlines or not.
 
 Fixes:
 ------